Statistics/pyproject.toml
[tool.poetry]
name = "statistics"
version = "0.1.0"
description = ""
authors = ["Nikolay Nekrasov <nnnekrasov@edu.hse.ru>"]
readme = "README.md"
packages = [{include = "statistics_service", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
pydantic-settings = "^2.2.1"
aiokafka = "^0.10.0"
sqlalchemy = "^2.0.29"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
pydantic = {extras = ["email"], version = "^2.7.1"}
pytest = "^8.2.2"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.7"
grpcio = "^1.64.1"
google = "^3.0.0"
grpcio-tools = "^1.64.1"
testcontainers = "^4.5.1"
greenlet = "^3.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
